requirements.txt
pandas>=1.3.0
openai>=1.0.0
loguru>=0.6.0
pydantic>=2.0.0
PyYAML>=6.0
tqdm>=4.62.0
numpy>=1.21.0
pyarrow
pyrootutils>=1.0.4
httpx>=0.24.0
jsonlines>=3.0.0
urllib3>=1.26.0